backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
yt-dlp==2026.3.17
moviepy==1.0.3
google-generativeai==0.7.2
google-api-python-client==2.134.0
google-auth-oauthlib==1.2.0
requests==2.32.3
supabase==2.5.0
python-multipart==0.0.9
apscheduler==3.11.2
pytz==2026.1.post1
